- #110: Moved default context in the file to fix cyclic dependency.
