--- src/app/hanzhong/utils/default-hanzhong-context-data.ts
@@ -1,10 +1,0 @@
-import type { HanzhongContextType } from '../types';
-
-export const DEFAULT_HANZHONG_CONTEXT_DATA: HanzhongContextType = {
-  user: {},
-  bonuses: {},
-  onChange: () => {},
-  formationsUserData: {},
-  onChangeFormations: () => {},
-  cache: { requirements: {} },
-};
