+++ src/app/hanzhong/utils/index.ts
@@ -1,6 +1,5 @@
 export { addHanzhongBonuses } from './add-hanzhong-bonuses';
 export { calculateSpecialTrainingsTacticalPoints } from './calculate-special-trainings-tactical-points';
-export { DEFAULT_HANZHONG_CONTEXT_DATA } from './default-hanzhong-context-data';
 export { findResourcesNeededByTechIdAndLevel } from './find-resources-needed-by-tech-id-and-level';
 export { getUserDataValueByKey } from './get-user-data-value-by-key';
 export { getUserDataValuesByKeys } from './get-user-data-values-by-keys';
